Resolve compression conflict
One of the /assets had a non Latin character(s) for the filename, so to avoid compression issues this file was renamed pre-commit. Then the file was renamed back to fit the codebase.

To avoid this conflict: modify the file /scripts/js.js and remove all non Latin characters. Then remove all non Latin characters from file in this commit.